Login Fixes
