+++ src/User.jsx
@@ -6,17 +6,18 @@ import * as ceramicApi from '../ceramic.js';
 // import styles from './User.module.css';
 import {storageHost, accountsHost, tokensHost, loginEndpoint} from '../constants';
 import {contracts, getAddressFromMnemonic} from '../blockchain.js';
-import {jsonParse, parseQuery} from '../util.js';
+import {jsonParse, parseQuery, handleDiscordLogin} from '../util.js';
 
 let userObject;
 
-function useComponentVisible(initialIsVisible) {
+function useComponentVisible(initialIsVisible, fn) {
   const [isComponentVisible, setIsComponentVisible] = useState(initialIsVisible);
   const ref = useRef(null);
 
   const handleClickOutside = event => {
     if (ref.current && !ref.current.contains(event.target)) {
       setIsComponentVisible(false);
+      fn();
     }
   };
 
@@ -30,83 +31,7 @@ function useComponentVisible(initialIsVisible) {
   return {ref, isComponentVisible, setIsComponentVisible};
 }
 
-async function contentIdToStorageUrl(id) {
-  if (typeof id === 'number') {
-    const hash = await contracts.mainnetsidechain.NFT.methods.getHash(id + '').call();
-    return `${storageHost}/${hash}`;
-  } else if (typeof id === 'string') {
-    return id;
-  } else {
-    return null;
-  }
-}
-
-async function pullUserObject(loginToken) {
-  const address = getAddressFromMnemonic(loginToken.mnemonic);
-  const res = await fetch(`${accountsHost}/${address}`);
-  const result = await res.json();
-  // console.log('pull user object', result);
-  let {name, avatarId, avatarName, avatarExt, avatarPreview, loadout, homeSpaceId, homeSpaceName, homeSpaceExt, homeSpaceFileName, homeSpacePreview, ftu} = result;
-  loadout = jsonParse(loadout, Array(8).fill(null));
-
-  const avatarNumber = parseInt(avatarId);
-  const avatarUrl = await contentIdToStorageUrl(!isNaN(avatarNumber) ? avatarNumber : avatarId);
-  const homeSpaceNumber = parseInt(homeSpaceId);
-  const homeSpaceUrl = await contentIdToStorageUrl(!isNaN(homeSpaceNumber) ? homeSpaceNumber : homeSpaceId);
-
-  const inventory = await (async () => {
-    const res = await fetch(`${tokensHost}/${address}`);
-    const tokens = await res.json();
-    return tokens;
-  })();
-
-  // menuActions.setInventory(inventory);
-
-  userObject = {
-    name,
-    avatar: {
-      id: avatarId,
-      name: avatarName,
-      ext: avatarExt,
-      preview: avatarPreview,
-      url: avatarUrl,
-    },
-    loadout,
-    inventory,
-    homespace: {
-      id: homeSpaceId,
-      name: homeSpaceName,
-      ext: homeSpaceExt,
-      preview: homeSpacePreview,
-      url: homeSpaceUrl,
-    },
-    ftu,
-  };
-  return userObject;
-}
-
-const handleDiscordLogin = async discordUrl => {
-  if (!discordUrl) {
-    return;
-  }
-  const urlSearchParams = new URLSearchParams(new URL(discordUrl).search);
-  const {id, code} = Object.fromEntries(urlSearchParams.entries());
-  let res = await fetch(loginEndpoint + `?discordid=${encodeURIComponent(id)}&discordcode=${encodeURIComponent(code)}`, {
-    method: 'POST',
-  });
-
-  res = await res.json();
-  if (!res.error) {
-    return await pullUserObject(res);
-  } else {
-    console.warn('Unable to login ', res.error);
-  }
-
-  // debugger;
-};
-
 const User = ({address, setAddress, open, setOpen, toggleOpen}) => {
-  const userRef = useComponentVisible(false);
   const discordRef = useComponentVisible(false);
   const metaMaskRef = useComponentVisible(false);
   const emailRef = useComponentVisible(false);
@@ -114,6 +39,41 @@ const User = ({address, setAddress, open, setOpen, toggleOpen}) => {
   const [loginButtons, setLoginButtons] = useState(false);
 
   const [discordUrl, setDiscordUrl] = useState('');
+  const [loginError, setLoginError] = useState(null);
+
+  const userRef = useComponentVisible(false, ()=>{
+    setLoginError(false);
+    setDiscordUrl('');
+  });
+
+  const metaMaskLogin = async e => {
+    e.preventDefault();
+    e.stopPropagation();
+    if (address) {
+      toggleOpen('user');
+    } else {
+      if (!loggingIn) {
+        setLoggingIn(true);
+        try {
+          const {address, profile} = await ceramicApi.login();
+          // console.log('login', {address, profile});
+          setAddress(address);
+        } catch (err) {
+          console.warn(err);
+        } finally {
+          setLoggingIn(false);
+        }
+      }
+    }
+  }
+
+  const discordLogin = async e => {
+    e.preventDefault();
+    e.stopPropagation();
+    setLoginButtons(false);
+    discordRef.setIsComponentVisible(true);
+  } 
+
 
   return (
     <div ref={userRef.ref}>
@@ -121,13 +81,17 @@ const User = ({address, setAddress, open, setOpen, toggleOpen}) => {
         onClick={async e => {
           e.preventDefault();
           e.stopPropagation();
-          userRef.setIsComponentVisible(true);
-          setLoginButtons(true);
-          discordRef.setIsComponentVisible(false);
-          metaMaskRef.setIsComponentVisible(false);
+          if (address) {
+            toggleOpen('user');
+          }else{
+            userRef.setIsComponentVisible(true);
+            setLoginButtons(true);
+            discordRef.setIsComponentVisible(false);
+            metaMaskRef.setIsComponentVisible(false);
+          }
         }}>
         <img src="images/soul.png" className={styles.icon} />
-        <div className={styles.name}>{loggingIn ? 'Logging in... ' : (address || 'Log in')}
+        <div className={styles.name}>{loggingIn ? 'Logging in... ' : (address || (loginError || 'Log in'))}
         </div>
       </div>
       {
@@ -135,37 +99,13 @@ const User = ({address, setAddress, open, setOpen, toggleOpen}) => {
           ? <div className={styles.login_options}>
             {
               loginButtons ? <>
-                <a className={styles.metamask} onClick={async e => {
-                  e.preventDefault();
-                  e.stopPropagation();
-                  if (address) {
-                    toggleOpen('user');
-                  } else {
-                    if (!loggingIn) {
-                      setLoggingIn(true);
-                      try {
-                        const {address, profile} = await ceramicApi.login();
-                        // console.log('login', {address, profile});
-                        setAddress(address);
-                      } catch (err) {
-                        console.warn(err);
-                      } finally {
-                        setLoggingIn(false);
-                      }
-                    }
-                  }
-                } }>
-                  <img src="https://upload.wikimedia.org/wikipedia/commons/thumb/3/36/MetaMask_Fox.svg/2048px-MetaMask_Fox.svg.png" width="30px" alt="" />
+                <a className={styles.metamask} onClick={ metaMaskLogin }>
+                  <img src="./images/metamask.png" width="30px" alt="" />
                   <span className={styles.metamask_text}
                   >MetaMask</span>
                 </a>
-                <a className={styles.discord} onClick={async e => {
-                  e.preventDefault();
-                  e.stopPropagation();
-                  setLoginButtons(false);
-                  discordRef.setIsComponentVisible(true);
-                } }>
-                  <img src="https://www.freepnglogos.com/uploads/discord-logo-png/concours-discord-cartes-voeux-fortnite-france-6.png" width="30px" alt="" />
+                <a className={styles.discord} onClick={discordLogin}>
+                  <img src="./images/discord.png" width="30px" alt="" />
                   <span className={styles.discord_text}>Discord</span>
                 </a></> : ''
             }
@@ -188,8 +128,11 @@ const User = ({address, setAddress, open, setOpen, toggleOpen}) => {
                                 event.preventDefault();
                                 event.stopPropagation();
                                 if (event.key === 'Enter') {
-                                  const {address} = await handleDiscordLogin(discordUrl);
-                                  setAddress(address);
+                                  const {address, error} = await handleDiscordLogin(discordUrl);
+                                  if(address)
+                                    setAddress(address);
+                                  else
+                                    setLoginError(String(error).toLocaleUpperCase());
                                 }
                               }
                             } />
